[time_based] Move TimeBasedActuatorBase from actuator/ to time_based/ component
- Move time_based_actuator.h/.cpp from actuator/ to time_based/
- Change namespace from esphome::actuator to esphome::time_based
- Move TIME_BASED_ACTUATOR_SCHEMA, apply_time_based_actuator_config,
  CONF_HAS_BUILT_IN_ENDSTOP, CONF_MANUAL_CONTROL from actuator/__init__.py
  to time_based/__init__.py
- Update time_based/cover/__init__.py to import from local package
- Update time_based/cover/time_based_cover.h to include local header
  and use unqualified TimeBasedActuatorBase (same namespace)
- Strip time-based helpers from actuator/__init__.py; actuator now only
  contains shared base infrastructure (ActuatorBase, ActuatorCallBase,
  IActuator, ActuatorOperation, build_apply_lambda_action)
