+++ esphome/components/time_based/__init__.py
@@ -1,3 +1,52 @@
+"""Time-based component — TimeBasedActuatorBase and shared schema."""
+
+from esphome import automation
 import esphome.codegen as cg
+import esphome.config_validation as cv
+from esphome.const import (
+    CONF_ASSUMED_STATE,
+    CONF_CLOSE_ACTION,
+    CONF_CLOSE_DURATION,
+    CONF_OPEN_ACTION,
+    CONF_OPEN_DURATION,
+    CONF_STOP_ACTION,
+)
+
+DEPENDENCIES = ["actuator"]
+
+CONF_HAS_BUILT_IN_ENDSTOP = "has_built_in_endstop"
+CONF_MANUAL_CONTROL = "manual_control"
 
 time_based_ns = cg.esphome_ns.namespace("time_based")
+
+TimeBasedActuatorBase = time_based_ns.class_("TimeBasedActuatorBase", cg.Component)
+
+TIME_BASED_ACTUATOR_SCHEMA = cv.Schema(
+    {
+        cv.Required(CONF_STOP_ACTION): automation.validate_automation(single=True),
+        cv.Required(CONF_OPEN_ACTION): automation.validate_automation(single=True),
+        cv.Required(CONF_OPEN_DURATION): cv.positive_time_period_milliseconds,
+        cv.Required(CONF_CLOSE_ACTION): automation.validate_automation(single=True),
+        cv.Required(CONF_CLOSE_DURATION): cv.positive_time_period_milliseconds,
+        cv.Optional(CONF_HAS_BUILT_IN_ENDSTOP, default=False): cv.boolean,
+        cv.Optional(CONF_MANUAL_CONTROL, default=False): cv.boolean,
+        cv.Optional(CONF_ASSUMED_STATE, default=True): cv.boolean,
+    }
+)
+
+
+async def apply_time_based_actuator_config(var, config):
+    await automation.build_automation(
+        var.get_stop_trigger(), [], config[CONF_STOP_ACTION]
+    )
+    cg.add(var.set_open_duration(config[CONF_OPEN_DURATION]))
+    await automation.build_automation(
+        var.get_open_trigger(), [], config[CONF_OPEN_ACTION]
+    )
+    cg.add(var.set_close_duration(config[CONF_CLOSE_DURATION]))
+    await automation.build_automation(
+        var.get_close_trigger(), [], config[CONF_CLOSE_ACTION]
+    )
+    cg.add(var.set_has_built_in_endstop(config[CONF_HAS_BUILT_IN_ENDSTOP]))
+    cg.add(var.set_manual_control(config[CONF_MANUAL_CONTROL]))
+    cg.add(var.set_assumed_state(config[CONF_ASSUMED_STATE]))
