+++ esphome/components/time_based/cover/time_based_cover.h
@@ -1,13 +1,13 @@
 #pragma once
 
-#include "esphome/components/actuator/time_based_actuator.h"
+#include "../time_based_actuator.h"
 #include "esphome/components/cover/cover.h"
 
 namespace esphome::time_based {
 
 // Inheritance: TimeBasedCover -> TimeBasedActuatorBase (Component), Cover (ActuatorBase, IActuator)
 // No shared ancestors. No diamond inheritance.
-class TimeBasedCover : public actuator::TimeBasedActuatorBase, public cover::Cover {
+class TimeBasedCover : public TimeBasedActuatorBase, public cover::Cover {
  public:
   TimeBasedCover() { this->set_actuator(this); }
   void setup() override;
@@ -15,7 +15,7 @@ class TimeBasedCover : public actuator::TimeBasedActuatorBase, public cover::Cov
   void dump_config() override;
 
  protected:
-  void control(const cover::CoverCall &call) override { actuator::TimeBasedActuatorBase::control(call); }
+  void control(const cover::CoverCall &call) override { TimeBasedActuatorBase::control(call); }
 };
 
 }  // namespace esphome::time_based
